refactor(merk/proofs): split aggregate_count.rs into aggregate_count/ subdirectory
Symmetric to the previous commit's aggregate_sum/ split. The
single-file `aggregate_count.rs` had grown to 2018 lines, mixing the
same six concerns as the sum side plus an extra `verify_only_tests`
module for the verify-only feature.

Split along the natural seams (parallel to aggregate_sum/):

| File                  | Lines | Contents                          |
|-----------------------|-------|-----------------------------------|
| `mod.rs`              |   76  | Module docs, layout doc-comment, public re-export of `verify_aggregate_count_on_range_proof`, the two small tree-type helpers `is_provable_count_bearing` / `provable_count_from_aggregate` |
| `prove.rs`            |   87  | `impl RefWalker { create_aggregate_count_on_range_proof, count_aggregate_on_range }` |
| `emit.rs`             |  263  | `emit_count_proof` — the recursive proof-emission engine |
| `walk.rs`             |  178  | `walk_count_only` + `provable_count_from_walker` (no-proof variant) |
| `verify.rs`           |  261  | `verify_aggregate_count_on_range_proof` + `verify_count_shape` |
| `tests.rs`            | 1166  | Main test module (prover + verifier integration tests, fuzz tests) |
| `verify_only_tests.rs`|  115  | Verify-only feature smoke tests (verifies pre-captured proof fixtures) |

Total: ~2146 lines across 7 files (was 2018 in one) — small growth
from per-file `//!` doc headers and explicit imports.

Behavior is identical:
- `cargo test --workspace`: 3199 / 0 fail (unchanged).
- `cargo clippy --workspace --all-features`: clean.
- The public surface (`Merk::prove_aggregate_count_on_range`,
  `Merk::count_aggregate_on_range`, `verify_aggregate_count_on_range_proof`)
  is untouched.

The `verify_only_tests` module sits in its own file (separate from
`tests.rs`) so the verify-only crate feature can build/exercise just
the verifier-only assertions, while `tests.rs` pulls in the full
`minimal` feature for prover-side machinery. The shared `FIXTURE_*`
constants are still `pub(super)`, so the drift-check test in
`tests.rs` continues to access them via `super::verify_only_tests::*`.
